sql: define flags for OP_MakeRecord
This patch defines three flags for OPcode OP_MakeRecord.

The first flag is OPFLAG_IS_EPHEMERAL allows to do some
optimization as created record will be inserted into ephemeral
space. This flag was created before this patch, but only here it
gotten its own name.

The second flag is OPFLAG_P2_IS_REG allows to use register P2 not
only as count of fields of created record, but also as number of
register that contains count of fields. This feature needed to
create records with number of fields determined during execution
of VDBE.

The third flag is OPFLAG_MAKE_ARRAY allows to change subtype of
created record to SQL_SUBTYPE_MSGPACK. Record, with this subtype
is considered as ARRAY and can be used as part of new record.

Needed for #2843