Do not latch schema while checkpoint
Increase a primary key indexes reference count until the end of a
checkpoint instead of schema latching. Attached test is not valid in
case of vinyl engine because of vy_log which is not able to stop
recovery on recoverable instance vclock. This allows to do not to stop
replication on ddl if there is a running checkpoint and enables
checkpoint during an ddl operation.

Prerequisites: #1254