sql: disallow identical samples in statistics
Before this patch it was possible that there were less rows of
statistic in _sql_stat4 table than it should be according to
number of samples created during analyze. It was this way because
some of rows in statistics were identical and were replaced
during inserting statistics in table _sql_stat4.

This patch disallows creation of identical rows of statistics
during analyze. After this patch number of statistics in
_sql_stat4 will be at least no less than it was before.

Needed for #2843