console: always use built-in string functions
A user can clobber such functions or remove the whole module from
globals, so we need to save then locally. When an error occurs in a
console code, tarantool exits and leave a terminal in a bad state.

The same changes are made for 'tap' module to ease testing of such
tarantool states.

Added libisatty.so with isatty() that always returns 1 to perform
testing of the local tarantool console.

Fixes #4154.