httpc: fix redirection on libcurl-7.30 and older
libcurl-7.30 and older have a problem with handling redirection. libcurl
postpones removing from an internal hash table and invoking
CURLMOPT_SOCKETFUNCTION callback to remove a closed socket. If timings
are so that a new socket is opened before the old one is fully removed
and they have the same file descriptors, CURLMOPT_SOCKETFUNCTION
callback will not be invoked for the new socket. See
curl_easy_sock_opt_cb() comment for more information.

The described situation will likely occur during redirection (say, at
'302 Found' HTTP response) when a request is machine-local. The request
will hang in the case.

The issue was fixed in libcurl-7.31: see the [discussion][1], the [bug
1248][2] and the [fix][3] (after that libcurl react on socket close
immediately). This commit provides a workaround for old libcurl
versions.

[1]: https://curl.haxx.se/mail/lib-2013-03/0045.html
[2]: https://sourceforge.net/p/curl/bugs/1248/
[3]: https://github.com/curl/curl/commit/88c5c63ffc3312a8c8471b48a44ec5f50420f2e3

Fixes #4180.