box: rename struct trigger to lua_trigger
Renamed struct trigger to struct lua_trigger because the 'trigger'
name is too general. Currently Tarantool has non-persistent
AppServer Lua triggers represented with 'trigger' structure so as
SQL-specific persistent trigger object named 'sql_trigger'.

In scope of this patchset we going to significantly rework
triggers subsystem and introduce a new base trigger class named
'trigger', so this name is be better to be released.

Needed for #4343