sql: rework CREATE TABLE rule in parser
SQL grammar used to construct the sql_trigger object or to
generate a related VDBE instructions in a two functions with
a shared context: sql_trigger_begin and sql_trigger_end.

It used to be an AST compile or VDBE code generation in according
to parser::is_parse_only flag state - that is messy.

This patch reworks SQL rule a bit to call a function once, when
the whole trigger context is assembled to call the first function
that constructs AST or the second to emit VDBE code.

Needed for #4343.