sql: introduce cache for prepared statemets
This patch introduces cache (as data structure) to handle prepared
statements and a set of interface functions (insert, delete, find) to
operate on it. Cache under the hood uses LRU eviction policy. It is
implemented as a hash table with string keys (which contain original SQL
statements) and prepared statements (struct sql_stmt which is an alias
for struct Vdbe) as values. To realise LRU strategy we maintain list of
nodes: head is the newest prepared statement, tail a candidate to be
evicted. Size of cache is regulated via box.cfg{sql_cache_size}
parameter. Cache is global to all sessions. To erase cache manually,
one can set its size to 0. Default cache size is assumed to be 5 Mb.

Part of #2592