FIXUP: msgpackffi.decode can now be assigned to buf.rpos
Rewritten the test case. There are several reasons to do so.

First, the previous test case implementation uses test:iscdata() which
is not sufficient to verify whether the const qualifier is applied to a
cdata type, because of using ffi.istype() under hood, which ignores
'const'.

Second, the previous test case was invalid, because of wrong order of
ffi.cast() arguments.

Third, we don't need a real ibuf object, ffi.cast(ctype, 'a string') is
enough. This allows to simplify test cases.

Fourth, the test case is rewritten in a declarative manner to reduce
code duplication. This also allows to expand it w/o many changes for
msgpack.decode() function in the following commit.

I left only msgpackffi.decode_unchecked() cases. A bit context is needed
to describe why.

msgpack.decode() accepts two arguments: a buffer and its size (or a
string and an offset, but it is out of scope here).
msgpack.decode_unchecked() accepts only a buffer (or, again, a string
and an offset) and does not verify whether we perform reads out of the
buffer bounds. See #2755 for the formal description of the API.

msgpackffi module contains only msgpackffi.decode_unchecked().
msgpackffi.decode() is just alias for decode_unchecked: it does not
verify a buffer bounds. AFAIU the alias was added to unify
testing code for msgpack and msgpackffi modules.

Our website has no documentation about msgpackffi. I don't sure whether
its API should be considered as public API. However if we'll make the
module public, we'll need to implement buffer bound check. And only then
state that the API is stable and can be used by users.

If we don't consider msgpackffi.decode() as public function for now, it
make sense to left it untested when a testing code is not unified
between msgpack and msgpackffi modules.

However I unified this testing code for both modules in the following
commit, so it will be tested anyway: but in the commit where it looks
reasonable.