test: fix flaky replication/gh-* tests
The tests didn't cleanup _cluster table. Autocleanup is turned on
only in > 1.10 versions. In 1.10 these tests failed, when were
launched in one test-run worker, because they still remembered
previous already deleted instances.

The patch makes cleanup in the end of these tests. Autocleanup
still works on master, but for the sake of similarity the manual
cleanup is done for all branches, including > 1.10.

Closes #4645