vinyl: add assertions for cache
Introduce several assertion to make sure that entries make up a
consistent chain. Moreover, in case it is not so (i.e. we are operating
on a broken entry) log entry itself and its neighbours. Note that debug
check may affect performance since we have to fetch previous/next
BPS tree elements to verify their links.