iproto: introduce IPROTO_INFO request
This is a draft commit.
This patch adds additional type of iproto request - IPROTO_INFO (key is
0x47).  It can be used to ask instance for some additional information
(e.g. box.info.listen). It is sent with a msgpack array of options that
user wants to get (unknown options will be ignored). The reply is a
msgpack map with known options (some requested options can be omitted).
The only one option that can be asked now is IPROTO_INFO_HOST_PORT == 0.

The main idea of the patch is to send additional request to replica in
order to find out its box.info.listen value. This functionality has not
been added yet.

Part of #3363