sql: remove unused DOUBLE to INTEGER conversion
This patch removes the unused DOUBLE to INTEGER conversion from OP_Seek*
opcodes. This conversion is not used due to changes in the ApplyType
opcode after #3809. The next few patches will introduce new rules for
converting numbers (not just DOUBLE to INTEGER), and the implicit
conversion within the ApplyType opcode will be disabled for this case.

Part of #4230