test: add corpus to be used with fuzzers
Fuzzing tools uses evolutionary algorithms. Supplying seed corpus consisting
of good sample inputs is one of the best ways to improve fuzz target’s
coverage. Patch adds a corpuses that can be used with existed fuzzers.
The name of each file in the corpus is the sha1 checksum of its contents.

Corpus with http headers was added from [1] and [2].

1. https://google.github.io/oss-fuzz/getting-started/new-project-guide/#seed-corpus
2. https://en.wikipedia.org/wiki/List_of_HTTP_header_fields
3. https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers

The libFuzzer allow to minimize corpus with help of `-merge` flag:
when 1 is passed, any corpus inputs from the 2nd, 3rd etc. corpus directories
that trigger new code coverage will be merged into the first corpus directory,
when 0 is passed an existed corpus will be minimized.

All provided corpuses in a patch were minimized.

Part of #1809