test: fix flaky app-tap/tarantoolctl.test.lua
The execution of tarantoolctl is failed on "tarantoolctl eval" with message:
"Control socket exists, but Tarantool doesn't listen on it" and exit
code: 2

To fix the issue added function tctl_wait_eval which tries to call eval command
several times and exits on eval return code 0. Delay between attempts is 0.01.