box: introduce summary RO flag
An instance is writable if box.cfg.read_only is false, and it is
not orphan. Update of the final read-only state of the instance
needs to fire read-only update triggers, and notify the engines.
These 2 flags were easy and cheap to check on each operation, and
the triggers were easy to use since both flags are stored and
updated inside box.cc.

We're going to make txn_limbo affect read-only state. This would be
useful, since a write on a node other than limbo owner results in an
error anyway.

It means a check for being read-only would look like this:

    is_ro || is_orphan || (!txn_limbo_is_empty() && limbo->owner_id !=
                                                    instance_id)

This is significantly slower.

The patch introduces a new flag is_ro_summary. The flag
incorporates all the read-only conditions into one flag. When some
subsystem may change read-only state of the instance, it needs to
call box_update_ro_summary(), and the function takes care of
updating the summary flag, running the triggers, and notifying the
engines.

The limbo will use this function each time its owner changes or it gets
emptied out.

Needed for #1146

(cherry picked from commit 31dc4fafb2e7e1e98c2864c869f636bec034f6d7)