test: flaky replication/bootstrap_leader.test.lua
Found on heavy loaded hosts the issue:

  --- replication/bootstrap_leader.result	Wed Apr 15 15:47:41 2020
  +++ var/020_replication/bootstrap_leader.result	Sun Sep  6 21:57:40 2020
  @@ -52,10 +52,3 @@
       test_run:cmd("stop server replica"..i.." with cleanup=1")
       test_run:cmd("delete server replica"..i)
   end;
  - | ---
  - | ...
  -
  -test_run:cmd("setopt delimiter ''");
  - | ---
  - | - true
  - | ...

It happened, because the test used static timeouts, which on heavy
loaded hosts where not tuned for it. To fix the issue the test must
be run in standalone mode by test-run tool. To make it available the
test was added to 'fragile' list of its suite configuration file.

Closes #5275